requirements.txt
fastapi
uvicorn
streamlit
opencv-python-headless
pytesseract
numpy
Pillow
presidio-analyzer
presidio-anonymizer
pymupdf
spacy
python-multipart
requests
